+++ src/py/bbctrl/Jog.py
@@ -80,7 +80,9 @@ class Jog(inevent.JogHandler):
                 axes = {}
                 for i in range(len(self.v)): axes["xyzabc"[i]] = self.v[i]
                 self.ctrl.mach.jog(axes)
-            except Exception as e: log.warning('Jog: %s', e)
+
+            except Exception as e:
+                log.warning('Jog: %s', e)
 
         self.ctrl.ioloop.call_later(0.25, self.callback)